# portals/common/core/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "apk-site Admin / Release / Ops API",
    "version": "2.0.0",
    "description": "Auto-generated from Flask url_map (admin portal mode)."
  },
  "paths": {
...
    "/admin": {
      "get": {
        "summary": "admin routes.admin panel",
        "operationId": "admin_routes.admin_panel_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/approval": {
      "get": {
        "summary": "admin routes.admin approval page",
        "operationId": "admin_routes.admin_approval_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/audit-log": {
      "get": {
        "summary": "admin routes.admin audit log page",
        "operationId": "admin_routes.admin_audit_log_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/my-tasks": {
      "get": {
        "summary": "admin routes.my tasks page",
        "operationId": "admin_routes.my_tasks_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/notifications": {
      "get": {
        "summary": "admin routes.admin notifications page",
        "operationId": "admin_routes.admin_notifications_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/projects": {
      "get": {
        "summary": "admin routes.admin projects page",
        "operationId": "admin_routes.admin_projects_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/projects/{project_id}/tasks": {
      "get": {
        "summary": "admin routes.project tasks page",
        "operationId": "admin_routes.project_tasks_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/projects/{project_id}/tasks/stats": {
      "get": {
        "summary": "admin routes.project tasks stats page",
        "operationId": "admin_routes.project_tasks_stats_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/reports": {
      "get": {
        "summary": "admin routes.admin reports page",
        "operationId": "admin_routes.admin_reports_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/settings": {
      "get": {
        "summary": "admin routes.admin settings page",
        "operationId": "admin_routes.admin_settings_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/site-config": {
      "get": {
        "summary": "admin routes.admin site config page",
        "operationId": "admin_routes.admin_site_config_page_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/site-config/editor/{portal_kind}": {
      "get": {
        "summary": "admin routes.admin site config visual editor",
        "operationId": "admin_routes.admin_site_config_visual_editor_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "portal_kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/release/bundles": {
      "get": {
        "summary": "release scopes.get release bundles",
        "operationId": "release_scopes.get_release_bundles_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/release/scopes/{scope_id}": {
      "get": {
        "summary": "release scopes.get release scope",
        "operationId": "release_scopes.get_release_scope_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/status": {
      "get": {
        "summary": "api.status",
        "operationId": "api.status_get",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}